updating Q class to use parser
